+++ server/queues/tasks/InviteReminderTask.ts
@@ -4,11 +4,13 @@ import { sequelize } from "@server/database/sequelize";
 import InviteReminderEmail from "@server/emails/templates/InviteReminderEmail";
 import { User } from "@server/models";
 import { UserFlag } from "@server/models/User";
-import BaseTask, { TaskPriority } from "./BaseTask";
+import BaseTask, { TaskPriority, TaskSchedule } from "./BaseTask";
 
 type Props = undefined;
 
 export default class InviteReminderTask extends BaseTask<Props> {
+  static cron = TaskSchedule.Daily;
+
   public async perform() {
     const users = await User.scope("invited").findAll({
       attributes: ["id"],
