+++ server/routes/api/index.ts
@@ -32,7 +32,6 @@ import subscriptions from "./subscriptions";
 import team from "./team";
 import users from "./users";
 import views from "./views";
-import webhookSubscriptions from "./webhookSubscriptions";
 
 const api = new Koa<AppState, AppContext>();
 const router = new Router();
@@ -82,7 +81,6 @@ router.use("/", attachments.routes());
 router.use("/", utils.routes());
 router.use("/", groups.routes());
 router.use("/", fileOperationsRoute.routes());
-router.use("/", webhookSubscriptions.routes());
 
 if (env.ENVIRONMENT === "development") {
   router.use("/", developer.routes());
