+++ shared/i18n/locales/en_US/translation.json
@@ -313,7 +313,6 @@
   "Shared Links": "Shared Links",
   "Import": "Import",
   "Integrations": "Integrations",
-  "Webhooks": "Webhooks",
   "Self Hosted": "Self Hosted",
   "Google Analytics": "Google Analytics",
   "Show path to document": "Show path to document",
@@ -671,22 +670,6 @@
   "Active": "Active",
   "Everyone": "Everyone",
   "Admins": "Admins",
-  "Are you sure you want to delete the {{ name }} webhook?": "Are you sure you want to delete the {{ name }} webhook?",
-  "Webhook updated": "Webhook updated",
-  "Update": "Update",
-  "Updating": "Updating",
-  "Provide a descriptive name for this webhook and the URL we should send a POST request to when matching events are created.": "Provide a descriptive name for this webhook and the URL we should send a POST request to when matching events are created.",
-  "A memorable identifer": "A memorable identifer",
-  "URL": "URL",
-  "Signing secret": "Signing secret",
-  "Subscribe to all events, groups, or individual events. We recommend only subscribing to the minimum amount of events that your application needs to function.": "Subscribe to all events, groups, or individual events. We recommend only subscribing to the minimum amount of events that your application needs to function.",
-  "All events": "All events",
-  "All {{ groupName }} events": "All {{ groupName }} events",
-  "Delete webhook": "Delete webhook",
-  "Disabled": "Disabled",
-  "Subscribed events": "Subscribed events",
-  "Edit webhook": "Edit webhook",
-  "Webhook created": "Webhook created",
   "Settings saved": "Settings saved",
   "Logo updated": "Logo updated",
   "Unable to upload new logo": "Unable to upload new logo",
@@ -770,6 +753,7 @@
   "Settings that impact the access, security, and content of your knowledge base.": "Settings that impact the access, security, and content of your knowledge base.",
   "Allow members to sign-in with {{ authProvider }}": "Allow members to sign-in with {{ authProvider }}",
   "Connected": "Connected",
+  "Disabled": "Disabled",
   "Allow members to sign-in using their email address": "Allow members to sign-in using their email address",
   "The server must have SMTP configured to enable this setting": "The server must have SMTP configured to enable this setting",
   "Access": "Access",
@@ -793,9 +777,6 @@
   "You can create an unlimited amount of personal tokens to authenticate\n          with the API. Tokens have the same permissions as your user account.\n          For more details see the <em>developer documentation</em>.": "You can create an unlimited amount of personal tokens to authenticate\n          with the API. Tokens have the same permissions as your user account.\n          For more details see the <em>developer documentation</em>.",
   "Tokens": "Tokens",
   "Create a token": "Create a token",
-  "New webhook": "New webhook",
-  "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.": "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.",
-  "Create a webhook": "Create a webhook",
   "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.": "Zapier is a platform that allows {{appName}} to easily integrate with thousands of other business tools. Automate your workflows, sync data, and more.",
   "Your are creating a new workspace using your current account — <em>{{email}}</em>": "Your are creating a new workspace using your current account — <em>{{email}}</em>",
   "Workspace name": "Workspace name",
@@ -838,5 +819,24 @@
   "Post to Channel": "Post to Channel",
   "This is what we found for \"{{ term }}\"": "This is what we found for \"{{ term }}\"",
   "No results for \"{{ term }}\"": "No results for \"{{ term }}\"",
+  "Are you sure you want to delete the {{ name }} webhook?": "Are you sure you want to delete the {{ name }} webhook?",
+  "Webhook updated": "Webhook updated",
+  "Update": "Update",
+  "Updating": "Updating",
+  "Provide a descriptive name for this webhook and the URL we should send a POST request to when matching events are created.": "Provide a descriptive name for this webhook and the URL we should send a POST request to when matching events are created.",
+  "A memorable identifer": "A memorable identifer",
+  "URL": "URL",
+  "Signing secret": "Signing secret",
+  "Subscribe to all events, groups, or individual events. We recommend only subscribing to the minimum amount of events that your application needs to function.": "Subscribe to all events, groups, or individual events. We recommend only subscribing to the minimum amount of events that your application needs to function.",
+  "All events": "All events",
+  "All {{ groupName }} events": "All {{ groupName }} events",
+  "Delete webhook": "Delete webhook",
+  "Subscribed events": "Subscribed events",
+  "Edit webhook": "Edit webhook",
+  "Webhook created": "Webhook created",
+  "Webhooks": "Webhooks",
+  "New webhook": "New webhook",
+  "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.": "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.",
+  "Create a webhook": "Create a webhook",
   "Uploading": "Uploading"
 }
